Fix conversion

+++ schemes/mceliece348864f/vec/pk_gen.c
@@ -119,7 +119,7 @@ static int mov_columns(uint64_t mat[][ ((SYS_N + 63) / 64) ], uint32_t * perm)
 	for (k = j+1; k < 64; k++)
 	{
 			d = perm[ row + j ] ^ perm[ row + k ];
-			d &= same_mask(k, ctz_list[j]);
+			d &= same_mask((uint16_t)k, (uint16_t)ctz_list[j]);
 			perm[ row + j ] ^= d;
 			perm[ row + k ] ^= d;
 	}
@@ -139,7 +139,7 @@ static int mov_columns(uint64_t mat[][ ((SYS_N + 63) / 64) ], uint32_t * perm)
 		for (k = j+1; k < 64; k++)
 		{
 			d = buf[ j ] ^ buf[ k ];
-			d &= same_mask(k, ctz_list[j]);
+			d &= same_mask((uint16_t)k, (uint16_t)ctz_list[j]);
 			buf[ j ] ^= d;
 			buf[ k ] ^= d;
 		}
